Fix(entity): added this reference

+++ core/src/main/java/io/github/spaceSurvivor/Movable.java
@@ -7,7 +7,7 @@ public abstract class Movable extends Entity {
 
     protected float speed;
 
-    public Movable(Texture texture, float posX, float posY, int sizeX, int sizeY, float speed) {
+    public Movable(Texture texture, float posX, float posY, float sizeX, float sizeY, float speed) {
         super(texture, posX, posY, sizeX, sizeY);
         this.speed = speed * Map.getUnitScale(); // pixels per second
     }
